requirements.txt
python-dotenv
Flask==3.0.2
flask-marshmallow==0.15.0
marshmallow==3.20.1
redis==5.0.3
Flask-Caching==2.1.0 # interactua con redis para almacenar en caché
tenacity==8.2.3  # Librería para reintentos cuando la api falla, se le setea los intentos y el tiempo entre cada reintento en segundos
flask_cors
saga_py==1.1.0
requests==2.32.3